+++ fabric/src/main/java/com/mrpg_lib/fabric/datagen/MRPGLibDatagen.java
@@ -1,0 +1,12 @@
+package com.mrpg_lib.fabric.datagen;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+
+public class MRPGLibDatagen  implements DataGeneratorEntrypoint {
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+        FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
+
+    }
+}